Modified the definition of ioff in MOInfo, MCSCF, and PSIMRCC because it lead to errors for large basis set (>400). (FAE)
git-svn-id: https://sirius.chem.vt.edu/svn/psi3/trunk@3979 29f243b5-bcef-41bb-bbe7-8e20262f3206

+++ src/lib/libmoinfo/moinfo_base.cc
@@ -86,9 +86,9 @@ void MOInfoBase::compute_number_of_electrons()
 
 void MOInfoBase::compute_ioff()
 {
-  ioff    = new int[IOFF];
+  ioff    = new size_t[IOFF];
   ioff[0] = 0;
-  for(unsigned long int i=1; i < IOFF; i++)
+  for(size_t i=1; i < IOFF; i++)
     ioff[i] = ioff[i-1] + i;
 }
